% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMplot_func.R
\name{KMplot.onlytwogrp}
\alias{KMplot.onlytwogrp}
\title{KM plot with only two subgroups}
\usage{
KMplot.onlytwogrp(
  dat,
  trt = "trt",
  tte,
  ttecen,
  group,
  strat = NA,
  ttime.inc = 4,
  switch.leg = TRUE,
  censor.status = FALSE,
  legend.type = 1,
  level.grp = c("Abema+EDT", "EDT"),
  legend.lb = c("Abema+EDT", "EDT"),
  legend.title.h = "Legend head",
  cols = c("red", "blue", "orange", "green"),
  legend.pos = c(0, 0),
  legend.adj = c(-0.1, -0.3),
  expand.val = c(0, 0),
  legend.title = 11,
  legend.text = 10,
  axis.title = 11,
  axis.text = 10,
  yname = "",
  margin.plot = c(1, 1, 1, 1),
  margin.tb = c(1, 1, 1, 1),
  xlim.lo = 0,
  xminlab = -2,
  xmaxlab = -1,
  heights.plot = c(1.7, 0.75),
  n.round = 2,
  p.round = 3,
  n.space = list(c(14, 5, 7, 15, 5, 7))
)
}
\arguments{
\item{dat}{data plot}

\item{trt}{treatments/arms}

\item{tte}{time to event variable}

\item{ttecen}{censor variable}

\item{group}{subgroup variable}

\item{strat}{whether using stratification}

\item{ttime.inc}{incremate interval time}

\item{switch.leg}{just for special case, in which legends switching}

\item{censor.status}{whether censor status displaying}

\item{legend.type}{for different legends format to display}

\item{level.grp}{subgroup levels}

\item{legend.lb}{legend labs}

\item{legend.title.h}{legend titles}

\item{cols}{KM curves colors}

\item{legend.pos}{legend position}

\item{legend.adj}{legend adjusted position}

\item{expand.val}{for expand x-axis}

\item{legend.title}{size of legend title}

\item{legend.text}{size of legend text}

\item{axis.title}{size of axis title}

\item{axis.text}{size of axis text}

\item{yname}{y-axis lab}

\item{margin.plot}{margins of plot}

\item{margin.tb}{margings of risk of table}

\item{xlim.lo}{x-axis limit}

\item{xminlab}{length of risk table legend}

\item{xmaxlab}{spaces between risk table legend and table}

\item{heights.plot}{the heights of plot}

\item{n.round}{statistics round number}

\item{p.round}{statistics round number}

\item{n.space}{the lists for adjusting spaces between values of legends}
}
\value{
KM plot with two curves
}
\description{
KM plot with only two subgroups
}
\examples{
# start example
attach(datplot)

library(survival)
library(survminer)
library(dplyr)
library(ggplot2)
library(gridExtra)
library(grid)

df.plot <- mutate(datplot, grp2=trt)

p1 <- KMplot.onlytwogrp(dat=df.plot,
                         tte="timevar",
                         ttecen="cenvar",
                         group = "grp2",
                         yname="Invasive Disease-Free Survival",
                         ttime.inc=3, legend.type=3,
                         legend.title.h="           Arm                 Median      HR (95\% CI)      p-value",
                         level.grp=c("Arm1", "Arm2"),
                         legend.lb=c('Arm1', 'Arm2'),
                         cols=c('red', "blue"),
                         legend.pos = c(0, 0),
                         legend.adj = c(-0.07, -0.3),
                         legend.title=11, legend.text=10,
                         axis.title=11, axis.text=10,
                         margin.plot=c(2, 2.5, 0, 1.5),
                         margin.tb=c(0, 2.1, 3, 0.7),
                         xminlab=-2.5, xmaxlab =-1, xlim.lo = -0.5,
                         heights.plot = c(2, 0.8),
                         n.space = list(c(17,16,17, 17,13)))
# different format
p2 <- KMplot.onlytwogrp(dat=df.plot,
                       tte="timevar",
                       ttecen="cenvar",
                       group = "grp2",
                       legend.type=1,
                       yname="Invasive Disease-Free Survival",
                       ttime.inc=3,
                       legend.title.h="           Arm                 Median      HR (95\% CI)",
                       level.grp=c("Arm1", "Arm2"),
                       legend.lb=c('Arm1', 'Arm2'),
                       cols=c('red', "blue"),
                       legend.pos = c(0, 0),
                       legend.adj = c(-0.07, -0.3),
                       legend.title=11, legend.text=10,
                       axis.title=11, axis.text=10,
                       margin.plot=c(2, 2.5, 0, 1.5),
                       margin.tb=c(0, 2.1, 3, 0.7),
                       xminlab=-2.5, xmaxlab =-1, xlim.lo = -0.5,
                       heights.plot = c(2, 0.8),
                       n.space = list(c(17, 16, 17, 13)))

# output
pdf(file=file.path(getwd(), "f_km_plot_2grp.pdf"), width=11, height=8.5)
grid.newpage()
print(p1, vp = viewport(x = unit(0.50, "npc"), y = unit(0.52, "npc"), width=0.90, height=0.85, just = "centre"))
grid.newpage()
print(p2, vp = viewport(x = unit(0.50, "npc"), y = unit(0.52, "npc"), width=0.90, height=0.85, just = "centre"))
dev.off()
#' # end example
}
