% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMplot_func.R
\name{KMplot.w.risk}
\alias{KMplot.w.risk}
\title{KM plot with more than two subgroups}
\usage{
KMplot.w.risk(
  dat,
  trt = "trt",
  tte,
  ttecen,
  group,
  strat = NA,
  ttime.inc = 4,
  switch.leg = FALSE,
  level.grp = NULL,
  subgrpName = c("grp1", "grp2", "grp3", "grp4"),
  subgrpComp = list(c("grp1", "grp2"), c("grp3", "grp4")),
  censor.status = FALSE,
  censor.size = 1,
  titles = "",
  legend.type = 1,
  legend.head = "Patient Group  HR (95\% CI)   P-value",
  legend.pos = c(0, 0),
  legend.adj = c(-0.05, -0.1),
  expand.val = c(0, 0),
  font.legend.txt = 10,
  font.legend.title = 11,
  color.val = c(1, 1, 1, 1),
  linetype = c("solid", "solid", "dashed", "dashed"),
  ylab.use = "",
  y.seq = seq(0, 10, 0.1),
  y.limit = c(0, 1),
  xlim.lo = 0,
  xminlab = -5,
  xmaxlab = -1,
  special.case = FALSE,
  hjust.val = 0.05,
  axis.text = 12,
  axis.title = 13,
  font.tb = 4,
  n.round = 2,
  p.round = 3,
  heights.plot = c(2, 0.85),
  margin.plot = c(1, 1, 0.1, 1.2),
  margin.tb = c(0.1, 0.2, 2, 2.1),
  n.space = list(c(5, 5, 5, 5, 5), c(5, 5, 5, 5, 5), c(5, 5, 5, 5, 5))
)
}
\arguments{
\item{dat}{data plot}

\item{trt}{treatments/arms}

\item{tte}{time to event variable}

\item{ttecen}{censor variable}

\item{group}{subgroup variable}

\item{strat}{whether using stratification}

\item{ttime.inc}{incremate interval time}

\item{switch.leg}{just for special case, in which legends switching}

\item{level.grp}{subgroup levels}

\item{subgrpName}{subgroup names}

\item{subgrpComp}{subgroups for comparison}

\item{censor.status}{whether censor status displaying}

\item{censor.size}{size of censor}

\item{titles}{title of plot}

\item{legend.type}{for different legends format to display}

\item{legend.head}{legend titles}

\item{legend.pos}{legend position}

\item{legend.adj}{adjusted legend position}

\item{expand.val}{for x-axis expanding}

\item{font.legend.txt}{size of legend text}

\item{font.legend.title}{size of legend title}

\item{color.val}{KM curves colors}

\item{linetype}{KM curve line types}

\item{ylab.use}{y-axis lab}

\item{y.seq}{y-axis sequence}

\item{y.limit}{y-axis limit}

\item{xlim.lo}{x-axis limit}

\item{xminlab}{length of risk table legend}

\item{xmaxlab}{spaces between risk table legend and table}

\item{special.case}{specical case for plotting risk lines legend}

\item{hjust.val}{spaces adjust for risk lines of legends if special.case is TRUE}

\item{axis.text}{size of axis text}

\item{axis.title}{size of axis title}

\item{font.tb}{size of text of table}

\item{n.round}{statistics round number}

\item{p.round}{statistics round number}

\item{heights.plot}{heights of plot}

\item{margin.plot}{margins of plot}

\item{margin.tb}{margins of table}

\item{n.space}{the lists for adjusting spaces between values of legends}
}
\value{
KM curves
}
\description{
KM plot with more than two subgroups
}
\examples{
# start example
attach(datplot)

library(survival)
library(survminer)
library(dplyr)
library(ggplot2)
library(gridExtra)
library(grid)

df.plot <- mutate(datplot2,
                  timevar=IDFS1_TIME, cenvar=IDFS1_CNSR,
                  grp=paste(grp.pos, trt, sep="_"),
                  trt=factor(trt, levels = c("EDT", "Abema+EDT"))) \%>\%
           select(-group)

p3 <- KMplot.w.risk(dat=df.plot,
                   trt = "trt",
                   tte="timevar", ttecen="cenvar", group="grp",
                   ttime.inc = 3,
                   subgrpName=c("ER+/PR+_Abema+EDT", "ER+/PR+_EDT",
                                "ER/PR_other_Abema+EDT", "ER/PR_other_EDT"),
                   subgrpComp=list(c("ER+/PR+_Abema+EDT", "ER+/PR+_EDT"),
                                   c("ER/PR_other_Abema+EDT", "ER/PR_other_EDT")),
                   censor.status = FALSE,
                   legend.type = 2,
                   legend.head="               Patient Group                          HR (95\% CI)     P-value",
                   legend.pos = c(0, 0),
                   legend.adj = c(-0.05, -0.1),
                   font.legend.txt = 10,
                   font.legend.title = 11,
                   color.val=c('red', 'blue', 'red', 'blue'),
                   linetype = c('solid','solid','dashed','dashed'),
                   #linetype=c(1,1,2,2),
                   ylab.use="Progression-Free Survival",
                   y.seq = seq(0, 10, 0.1),
                   y.limit = c(0, 1),
                   xminlab = -5,
                   xmaxlab = -2.2,
                   xlim.lo = -1,
                   axis.text = 12, axis.title = 13,
                   font.tb = 4,
                   n.round=3, p.round = 4,
                   special.case = TRUE,
                   heights.plot = c(2, 0.85),
                   margin.plot=c(1, 1, 0, 1),
                   margin.tb=c(0, 0.4, 2, 0.7),
                   n.space=list(c(20, 15, 29),
                                c(13, 15, 24)))
# output
pdf(file=file.path(getwd(), "f_km_plot_pair.comp.pdf"), width=11, height=8.5)
grid.newpage()
print(p3, vp = viewport(x = unit(0.52, "npc"), y = unit(0.54, "npc"), width=0.90, height=0.9, just = "centre"))
dev.off()
#' # end example
}
